# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development — بيئة التطوير"
    },
    {
      "url": "https://akfeek-backend.developteam.site",
      "description": "Production — بيئة الإنتاج"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token — أدخل التوكن بعد كلمة Bearer مثال: `Bearer eyJhbGci...`"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          },
          "errorAr": {
            "type": "string",
            "example": "رسالة الخطأ"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    },
    "parameters": {
...
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "رقم الصفحة | Page number",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "عدد العناصر في الصفحة | Items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "مطلوب مصادقة | Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": "Authentication required",
              "errorAr": "مطلوب مصادقة",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "تم رفض الوصول | Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": "Access denied",
              "errorAr": "تم رفض الوصول",
              "code": "FORBIDDEN"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "🚗 Akfeek API — Full Reference",
    "version": "1.0.0",
    "description": "\n# Akfeek Platform — Full API Reference\n\n> **للمطورين الداخليين فقط** — هذه الوثيقة تحتوي على **جميع** الـ endpoints.\n> لو بتشتغل على جزء معين، استخدم الرابط المخصص ليك:\n\n| الجمهور | الرابط |\n|---|---|\n| 📱 تطبيق موبايل العميل | [`/api-docs/mobile`](/api-docs/mobile) |\n| 🔧 تطبيق موبايل الفني | [`/api-docs/technician`](/api-docs/technician) |\n| 🏪 بوابة الفيندور / الورشة | [`/api-docs/vendor`](/api-docs/vendor) |\n| ⚙️ لوحة تحكم الأدمن | [`/api-docs/admin`](/api-docs/admin) |\n\n---\n\n## Authentication / المصادقة\n\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n1. `POST /api/auth/login` — سجل دخول واحصل على التوكن\n2. انقر **Authorize** في الأعلى وأدخل: `Bearer <token>`\n      ",
    "contact": {
      "name": "Akfeek API Support",
      "email": "support@akfeek.com"
    }
  },
...
  "paths": {
...
    "/api/bookings": {
      "get": {
        "summary": "List all bookings (Admin)",
        "description": "Retrieve a paginated list of all bookings in the system.",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings"
          }
        }
      },
      "post": {
        "summary": "Create new booking",
        "description": "Create a new booking with optional workshop and delivery method",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "vehicleId",
                  "scheduledDate"
                ],
                "properties": {
                  "customerId": {
                    "type": "string"
                  },
                  "vehicleId": {
                    "type": "string"
                  },
                  "scheduledDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "scheduledTime": {
                    "type": "string",
                    "example": "10:00"
                  },
                  "workshopId": {
                    "type": "string",
                    "description": "ID of certified workshop (optional)"
                  },
                  "deliveryMethod": {
                    "type": "string",
                    "enum": [
                      "FLATBED",
                      "SELF_DELIVERY"
                    ],
                    "description": "Required if workshopId is provided"
                  },
                  "serviceIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "productIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully"
          },
          "400": {
            "description": "Invalid input or workshop not available"
          },
          "404": {
            "description": "Workshop not found"
          }
        }
      }
    },
    "/api/bookings/my": {
      "get": {
        "summary": "Get my bookings",
        "description": "Retrieve bookings for the authenticated customer.",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user bookings"
          }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get booking by ID",
        "description": "Retrieve details of a specific booking.",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/api/bookings/{bookingId}/track": {
      "get": {
        "summary": "Track technician location",
        "description": "Customer tracks technician's real-time location",
        "tags": [
          "🔧 Technician | Location"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bookingId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tracking information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "booking": {
                          "type": "object"
                        },
                        "technician": {
                          "type": "object",
                          "properties": {
                            "currentLocation": {
                              "type": "object",
                              "properties": {
                                "latitude": {
                                  "type": "number"
                                },
                                "longitude": {
                                  "type": "number"
                                },
                                "heading": {
                                  "type": "number"
                                },
                                "speed": {
                                  "type": "number"
                                }
                              }
                            }
                          }
                        },
                        "eta": {
                          "type": "object",
                          "properties": {
                            "distanceKm": {
                              "type": "number"
                            },
                            "durationMinutes": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/status": {
      "patch": {
        "summary": "Update booking status",
        "description": "Update booking status. Auto-creates invoice when status is COMPLETED or DELIVERED.",
        "tags": [
          "⚙️ Admin | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "CONFIRMED",
                      "IN_PROGRESS",
                      "COMPLETED",
                      "DELIVERED",
                      "CANCELLED",
                      "TECHNICIAN_EN_ROUTE",
                      "ARRIVED",
                      "IN_SERVICE"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  },
                  "technicianId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated. Invoice auto-created if COMPLETED/DELIVERED."
          }
        }
      }
    },
    "/api/bookings/{bookingId}/location-history": {
      "get": {
        "summary": "Get location history",
        "description": "Retrieve technician's route history for completed job",
        "tags": [
          "🔧 Technician | Location"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bookingId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "startTime",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endTime",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location history points"
          }
        }
      }
    },
...
    "/api/mobile-car-service/bookings": {
      "post": {
        "summary": "Create mobile car service booking",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "vehicleId",
                  "scheduledDate",
                  "serviceIds"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created"
          }
        }
      }
    },
    "/api/mobile-car-service/bookings/{id}": {
      "get": {
        "summary": "Get mobile service booking details",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details"
          }
        }
      }
    },
    "/api/mobile-car-service/bookings/{id}/status": {
      "patch": {
        "summary": "Update mobile service booking status",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        }
      }
    },
...
    "/api/technician/bookings": {
      "get": {
        "summary": "Get my assigned bookings (Technician)",
        "description": "الحجوزات العامة المعينة للفني (غير الونش والغسيل)\nReturns general bookings assigned to the authenticated technician.\n",
        "tags": [
          "🔧 Technician | My Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by booking status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of assigned bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
    "/api/workshops/profile/me/bookings": {
      "get": {
        "summary": "Get my workshop bookings",
        "tags": [
          "🏪 Vendor | Workshop"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of workshop bookings"
          }
        }
      }
    },
...
  }
}